Move model file.
