+++ README.md
@@ -4,3 +4,27 @@
 docker-compose up -d
 ```
 ## Project Structure
+Layered Architecture(分層式架構)
+```markdown
+├── api
+│   ├── handler
+│   ├── middleware
+│   ├── request
+│   ├── response
+│   ├── model
+│   ├── routes
+│   ├── utils
+│
+├── internal
+│   ├── repository
+│   ├── service
+│
+├── docs
+│
+├── go.mod
+├── go.sum
+│
+├── test
+│
+├── main.go
+```
