+++ internal/service/user/signup.go
@@ -2,11 +2,11 @@ package user
 
 import (
 	"errors"
-	"study_savvy_api_go/api/model"
 	"study_savvy_api_go/api/request/user"
 	responseUser "study_savvy_api_go/api/response/user"
 	responseUtils "study_savvy_api_go/api/response/utils"
 	"study_savvy_api_go/api/utils"
+	"study_savvy_api_go/internal/repository/model"
 	"study_savvy_api_go/internal/repository/sql"
 	StatusUtils "study_savvy_api_go/internal/repository/utils"
 )
